+++ src/developer-platform/contracts/future-developer-platform-evolution-profile.js
@@ -1,0 +1,48 @@
+export class FutureDeveloperPlatformEvolutionProfile {
+  constructor({
+    evolutionName,
+    objectives = [], strategicDrivers = [], invariants = [], horizons = [], capabilityStates = [],
+    maturityDimensions = [], experienceAttributes = [], intentResolutionOutputs = [],
+    architectureAssistanceCapabilities = [], knowledgeGraphNodes = [], humanAiTeamRoles = [],
+    agentCoordinationFields = [], autonomyLevels = [], autonomyPromotionCriteria = [],
+    autonomyDemotionTriggers = [], futureCapabilityAreas = [], digitalTwinTypes = [], federationRequirements = [],
+    ecosystemEvolutionAreas = [], innovationControlFields = [], transitionArchitectureFields = [],
+    governanceAreas = [], successMeasures = [], risks = [], safeguards = [], antiPatterns = [],
+    architecturalRules = [], recommendedSequence = [],
+    preservesAdd = true, evidenceBackedEvolution = true, futureCapabilitiesOptionalUntilProven = true,
+    architectureRepositoryAuthoritative = true, intentDoesNotInventAuthority = true, semanticInferenceDistinguished = true,
+    explicitAgentCoordination = true, autonomyExplicitlyLeveled = true, autonomyPromotionEvidenceBased = true,
+    autonomyDemotionImmediateWhenUnsafe = true, digitalTwinsDecisionAidOnly = true, federationUsesCommonContracts = true,
+    controlledInnovation = true, transitionArchitectureRequired = true, backwardCompatibilityGoverned = true,
+    vendorNeutralAiProviderIndependent = true, humanAccountabilityPreserved = true, tenantPropertyIsolationPreserved = true,
+    preselectsVendors = false, mandatesSpeculativeTechnology = false, authorizesAutonomyWithoutControls = false,
+    treatsAiCapabilityAsAuthority = false, writesInferenceAsFact = false, treatsDigitalTwinAsOperatingState = false,
+    permitsUnboundedMultiAgentAuthority = false, permitsProductionDataInExperiments = false,
+    allowsPermanentExperimentalServices = false, letsVendorContractsBecomeArchitecture = false,
+    measuresNoveltyAsValue = false, erodesHumanExpertise = false, fragmentsFederatedAuthority = false
+  }) {
+    this.evolutionName = evolutionName;
+    for (const [key, value] of Object.entries({ objectives, strategicDrivers, invariants, horizons, capabilityStates,
+      maturityDimensions, experienceAttributes, intentResolutionOutputs, architectureAssistanceCapabilities,
+      knowledgeGraphNodes, humanAiTeamRoles, agentCoordinationFields, autonomyLevels, autonomyPromotionCriteria,
+      autonomyDemotionTriggers, futureCapabilityAreas, digitalTwinTypes, federationRequirements,
+      ecosystemEvolutionAreas, innovationControlFields, transitionArchitectureFields, governanceAreas, successMeasures,
+      risks, safeguards, antiPatterns, architecturalRules, recommendedSequence })) {
+      this[key] = Object.freeze([...value]);
+    }
+    for (const [key, value] of Object.entries({ preservesAdd, evidenceBackedEvolution,
+      futureCapabilitiesOptionalUntilProven, architectureRepositoryAuthoritative, intentDoesNotInventAuthority,
+      semanticInferenceDistinguished, explicitAgentCoordination, autonomyExplicitlyLeveled,
+      autonomyPromotionEvidenceBased, autonomyDemotionImmediateWhenUnsafe, digitalTwinsDecisionAidOnly,
+      federationUsesCommonContracts, controlledInnovation, transitionArchitectureRequired,
+      backwardCompatibilityGoverned, vendorNeutralAiProviderIndependent, humanAccountabilityPreserved,
+      tenantPropertyIsolationPreserved, preselectsVendors, mandatesSpeculativeTechnology,
+      authorizesAutonomyWithoutControls, treatsAiCapabilityAsAuthority, writesInferenceAsFact,
+      treatsDigitalTwinAsOperatingState, permitsUnboundedMultiAgentAuthority, permitsProductionDataInExperiments,
+      allowsPermanentExperimentalServices, letsVendorContractsBecomeArchitecture, measuresNoveltyAsValue,
+      erodesHumanExpertise, fragmentsFederatedAuthority })) {
+      this[key] = Boolean(value);
+    }
+    Object.freeze(this);
+  }
+}